refactor: ReservationTime 패키지 분리
