+++ src/main/java/roomescape/domain/reservation/repository/ReservationJdbcRepository.java
@@ -11,7 +11,7 @@ import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.domain.reservation.entity.Reservation;
-import roomescape.domain.reservation.entity.ReservationTime;
+import roomescape.domain.time.entity.ReservationTime;
 import roomescape.domain.theme.entity.Theme;
 
 @Repository
