+++ src/main/java/roomescape/domain/time/controller/ReservationTimeController.java
@@ -1,13 +1,13 @@
-package roomescape.domain.reservation.controller;
+package roomescape.domain.time.controller;
 
 import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.domain.reservation.response.ReservationTimeResponse;
-import roomescape.domain.reservation.response.ReservationTimesResponse;
-import roomescape.domain.reservation.service.ReservationTimeService;
+import roomescape.domain.time.response.ReservationTimeResponse;
+import roomescape.domain.time.response.ReservationTimesResponse;
+import roomescape.domain.time.service.ReservationTimeService;
 
 @RestController
 @RequestMapping("/times")
