+++ src/main/java/roomescape/domain/time/exception/ReservationTimeDeleteConflictException.java
@@ -1,4 +1,4 @@
-package roomescape.domain.reservation.exception;
+package roomescape.domain.time.exception;
 
 import roomescape.common.exception.BusinessException;
 import roomescape.common.exception.ErrorCode;
