+++ src/main/java/roomescape/domain/time/repository/ReservationTimeRepository.java
@@ -1,8 +1,8 @@
-package roomescape.domain.reservation.repository;
+package roomescape.domain.time.repository;
 
 import java.util.List;
 import java.util.Optional;
-import roomescape.domain.reservation.entity.ReservationTime;
+import roomescape.domain.time.entity.ReservationTime;
 
 public interface ReservationTimeRepository {
